requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.2
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
chromedriver==2.24.1
colorama==0.4.3
cryptography==3.4.6
cssutils==1.0.2
currency-symbols==1.0.0
cycler==0.10.0
distlib==0.3.1
Django==2.2
django-crispy-forms==1.10.0
django-filter==2.4.0
django-phone-field==1.8.1
django-resized==0.3.11
django-rest-framework==0.1.0
django-restframework==0.0.1
django-smartfields==1.1.3
djangorestframework==3.12.2
et-xmlfile==1.0.1
excel2img==1.4.0
filelock==3.0.12
fonttools==4.16.1
future==0.18.2
fuzzywuzzy==0.18.0
idna==2.10
isort==5.4.2
jdcal==1.4.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lml==0.1.0
lxml==4.5.2
Markdown==3.3.3
mccabe==0.6.1
pefile==2019.4.18
Pillow==8.1.0
progressbar==2.5
psutil==5.7.3
psycopg2==2.8.6
py2exe==0.10.1.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
PyMsgBox==1.0.9
pyparsing==2.4.7
PyPDF2==1.26.0
pypiwin32==223
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
reportlab==3.5.56
requests==2.24.0
six==1.15.0
sortedcontainers==2.3.0
soupsieve==2.0.1
sqlparse==0.4.1
texttable==1.6.3
toml==0.10.1
ttkthemes==3.2.0
twilio==6.48.0
urllib3==1.25.10
virtualenv==20.0.31
webcolors==1.11.1
wrapt==1.12.1